requirements.txt
torch
sentence-transformers
numpy
termcolor
pytest
rank_bm25
faiss-cpu
pdfplumber